+++ model/point.hpp
@@ -9,8 +9,8 @@
  */
 class Point
 {
-    int x {0};
-    int y {0};
+    double x {0.};
+    double y {0.};
 
 public:
 
@@ -24,45 +24,45 @@ public:
      * @param x l'abscisse du point
      * @param y l'ordonnée du point
      */
-    Point(const int, const int);
+    Point(const double, const double);
 
     /**
      * Retourne l'abscisse du point.
      * @return l'abscisse du point.
      */
-    int getX() const;
+    double getX() const;
 
     /**
      * Retourne l'ordonnée du point.
      * @return l'ordonnée du point.
      */
-    int getY() const;
+    double getY() const;
 
     /**
      * Déplace le point en l'abscisse donnée.
      * @param x l'abscisse où déplacer le point.
      */
-    void setX(const int);
+    void setX(const double);
 
     /**
      * Déplace le point en l'ordonnée donnée.
      * @param y l'ordonnée où déplacer le point.
      */
-    void setY(const int);
+    void setY(const double);
 
     /**
      * Déplace le point en la coordonnée donnée.
      * @param x l'abscisse où déplacer le point.
      * @param y l'ordonnée où déplacer le point.
      */
-    void setLocation(const int, const int);
+    void setLocation(const double, const double);
 
     /**
      * @brief distanceFrom
      * @param p
      * @return
      */
-    int distanceFrom(const Point & p);
+    double distanceFrom(const Point & p);
 
     /**
      * Permet de savoir si deux points sont aux même endroit.
